libmemcached: fix a reallocation crash
Also, re-enable memcached_server_add(many weighted) test which was
presumably disabled because of this issue.

When a continuum was allocated while not in weighted distribution mode,
the actual space allocated at first would be to small to fit further
updates when weighted changed to true at a later time.